requirements.txt
# Core Dependencies
torch>=2.1.0
pytorch-lightning>=2.1.0
torchvision>=0.16.0
torch-geometric>=2.4.0

# NLP and Phonetics
pythainlp>=4.0.0
epitran>=1.24
panphon>=0.20
python-crfsuite>=0.9.9

# Data Processing
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<1.25.0
datasets>=2.10.0
pyarrow>=12.0.0

# Graph Processing
networkx>=3.1
matplotlib>=3.5.0,<3.8.0
seaborn>=0.11.0

# Visualization
plotly>=5.17.0
dash>=2.14.0
dash-cytoscape>=0.3.0
kaleido>=0.2.1

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.5.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Experiments and Logging
wandb>=0.16.0
tensorboard>=2.15.0
mlflow>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
